Factor out clean-cast-synthesis code, and use it in java_bytecode_typecheck to turn casts-via-void into clean typesafe operations where possible.
