+++ src/java_bytecode/Makefile
@@ -5,7 +5,8 @@ SRC = java_bytecode_language.cpp java_bytecode_parse_tree.cpp \
       java_bytecode_typecheck_type.cpp java_bytecode_internal_additions.cpp \
       java_class_identifier.cpp java_bytecode_parser.cpp bytecode_info.cpp \
       java_class_loader.cpp jar_file.cpp java_object_factory.cpp \
-      java_opaque_method_stubs.cpp java_bytecode_convert_method.cpp
+      java_opaque_method_stubs.cpp java_bytecode_convert_method.cpp \
+      java_pointer_casts.cpp
 
 INCLUDES= -I ..
