+++ src/java_bytecode/java_bytecode_typecheck_expr.cpp
@@ -10,6 +10,7 @@ Author: Daniel Kroening, kroening@kroening.com
 #include <util/prefix.h>
 
 #include "java_bytecode_typecheck.h"
+#include "java_pointer_casts.h"
 
 /*******************************************************************\
 
@@ -27,6 +28,9 @@ void java_bytecode_typecheckt::typecheck_expr(exprt &expr)
 { 
   if(expr.id()==ID_code)
     return typecheck_code(to_code(expr));
+
+  if(expr.id()==ID_typecast && expr.type().id()==ID_pointer)
+    expr=make_clean_pointer_cast(expr,expr.type(),ns);
 
   // do operands recursively
   Forall_operands(it, expr)
