+++ src/java_bytecode/java_opaque_method_stubs.cpp
@@ -7,80 +7,10 @@
 #include <util/prefix.h>
 
 #include "java_object_factory.h"
+#include "java_pointer_casts.h"
 
 namespace
 { // Anon namespace for insert-nondet support functions
-
-exprt clean_deref(const exprt ptr)
-{
-
-  return ptr.id()==ID_address_of
-             ? ptr.op0()
-             : dereference_exprt(ptr,ptr.type().subtype());
-}
-
-bool find_superclass_with_type(exprt &ptr,const typet &target_type,
-                               const namespacet &ns)
-{
-
-  while(true)
-  {
-
-    const typet ptr_base=ns.follow(ptr.type().subtype());
-
-    if(ptr_base.id()!=ID_struct)
-      return false;
-
-    const struct_typet &base_struct=to_struct_type(ptr_base);
-
-    if(base_struct.components().size()==0)
-      return false;
-
-    const typet &first_field_type=
-        ns.follow(base_struct.components()[0].type());
-    ptr=clean_deref(ptr);
-    ptr=member_exprt(ptr,base_struct.components()[0].get_name(),
-                      first_field_type);
-    ptr=address_of_exprt(ptr);
-
-    if(first_field_type==target_type)
-      return true;
-  }
-}
-
-exprt make_clean_pointer_cast(const exprt ptr,const typet &target_type,
-                              const namespacet &ns)
-{
-
-  assert(target_type.id()==ID_pointer &&
-         "Non-pointer target in make_clean_pointer_cast?");
-
-  if(ptr.type()==target_type)
-    return ptr;
-
-  const typet &target_base=ns.follow(target_type.subtype());
-
-  exprt bare_ptr=ptr;
-  while(bare_ptr.id()==ID_typecast)
-  {
-    assert(bare_ptr.type().id()==ID_pointer &&
-           "Non-pointer in make_clean_pointer_cast?");
-    if(bare_ptr.type().subtype()==empty_typet())
-      bare_ptr=bare_ptr.op0();
-  }
-
-  assert(bare_ptr.type().id()==ID_pointer &&
-         "Non-pointer in make_clean_pointer_cast?");
-
-  if(bare_ptr.type()==target_type)
-    return bare_ptr;
-
-  exprt superclass_ptr=bare_ptr;
-  if(find_superclass_with_type(superclass_ptr,target_base,ns))
-    return superclass_ptr;
-
-  return typecast_exprt(bare_ptr,target_type);
-}
 
 void insert_nondet_opaque_fields_at(const typet &expected_type,
                                     const exprt &ptr,
